added scenarios to test for service and repo account and fixed others

+++ ServerApp/src/main/java/com/example/app/serviceImpl/AccountServiceImpl.java
@@ -1,8 +1,11 @@
 package com.example.app.serviceImpl;
 
+import java.util.NoSuchElementException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.example.app.exception.AccountNotFoundException;
 import com.example.app.exception.InsuficientFundsException;
 import com.example.app.exception.OverdrawnAccountException;
 import com.example.app.model.Account;
@@ -13,6 +16,8 @@ import com.example.app.service.AccountService;
 
 @Service
 public class AccountServiceImpl implements AccountService {
+	
+	private final String NOT_FOUND = "Account not found by id ";
 
 	@Autowired
 	private AccountRepository accountRepository;
@@ -26,45 +31,61 @@ public class AccountServiceImpl implements AccountService {
 	}
 	
 	@Override
-	public String deleteAccount(Long id) throws OverdrawnAccountException {
-
+	public String deleteAccount(Long id) throws OverdrawnAccountException, AccountNotFoundException {
 			Account lockedAccount = accountRepository.findAndLockById(id);
 			
-			if(lockedAccount.getBalance() >= 0) {
-				accountRepository.delete(lockedAccount);
-				return "Account closed!";
+			if(lockedAccount != null) {
+				if(lockedAccount.getBalance() >= 0) {
+					accountRepository.delete(lockedAccount);
+					return "Account closed!";
+				}
+				else {
+					throw new OverdrawnAccountException("Account's balance negative, unable to close");
+				}
+			} else {
+				throw new AccountNotFoundException(NOT_FOUND + id);
 			}
-			else {
-				throw new OverdrawnAccountException("Account's balance negative, unable to close");
-			}
 	}
 	
 	@Override
-	public double getBalance(Long id) {
-		return accountRepository.findById(id).get().getBalance();
+	public double getBalance(Long id) throws AccountNotFoundException  {
+		try {
+			return accountRepository.findById(id).get().getBalance();
+		} catch(NoSuchElementException e) {
+			throw new AccountNotFoundException(NOT_FOUND + id);
+		}
 	}
 	
 	@Override
-	public Account getAccountbyIdNPin(Long id, Integer pin) {
-		return accountRepository.findByIdAndPin(id, pin);
+	public Account getAccountbyIdNPin(Long id, Integer pin) throws AccountNotFoundException {
+		Account account = accountRepository.findByIdAndPin(id, pin);
+		if(account != null) {
+			return account;
+		} else {
+			throw new AccountNotFoundException(NOT_FOUND + id + " and PIN: " + pin);
+		}
 	}
 	
 	@Override
-	public Account save(Transaction transaction) throws InsuficientFundsException  {
+	public Account save(Transaction transaction) throws InsuficientFundsException, AccountNotFoundException  {
 		Account lockedAccount = accountRepository.findAndLockById(transaction.getAccount().getId());
 		
-		if(transaction.getType() == DEPOSIT) {
-			transactionRepository.save(transaction);
-			lockedAccount.setBalance(lockedAccount.getBalance() + transaction.getAmount());
-			return lockedAccount;
-		} 
-		else if(lockedAccount.getBalance() >= transaction.getAmount()) {
-			transactionRepository.save(transaction);
-			lockedAccount.setBalance(lockedAccount.getBalance() - transaction.getAmount());
-			return lockedAccount;
-		} 
-		else { 
-			throw new InsuficientFundsException("operation unsuccessful, insuficient funds" + lockedAccount.getId());
+		if(lockedAccount != null ) {
+			if(transaction.getType() == DEPOSIT) {
+				transactionRepository.save(transaction);
+				lockedAccount.setBalance(lockedAccount.getBalance() + transaction.getAmount());
+				return lockedAccount;
+			} 
+			else if(lockedAccount.getBalance() >= transaction.getAmount()) {
+				transactionRepository.save(transaction);
+				lockedAccount.setBalance(lockedAccount.getBalance() - transaction.getAmount());
+				return lockedAccount;
+			} 
+			else { 
+				throw new InsuficientFundsException("operation unsuccessful, insuficient funds" + lockedAccount.getId());
+			}
+		} else {
+			throw new AccountNotFoundException(NOT_FOUND + transaction.getAccount().getId());
 		}
 	}
 }
